Refactor: 맛집 API fetch 커스텀훅 분리 및 적용
